+++ server/src/main/java/com/blocklang/core/dao/PropertyDao.java
@@ -1,0 +1,13 @@
+package com.blocklang.core.dao;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.blocklang.core.model.CmProperty;
+
+public interface PropertyDao  extends JpaRepository<CmProperty, Integer>{
+
+	Optional<CmProperty> findByKeyAndParentIdAndValid(String key, Integer parentId, boolean valid);
+
+}