+++ server/src/main/java/com/blocklang/core/service/impl/PropertyServieImpl.java
@@ -1,0 +1,26 @@
+package com.blocklang.core.service.impl;
+
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.blocklang.core.dao.PropertyDao;
+import com.blocklang.core.model.CmProperty;
+import com.blocklang.core.service.PropertyService;
+
+@Service
+public class PropertyServieImpl implements PropertyService {
+
+	private static final int ROOT_ID = -1;
+	
+	@Autowired
+	public PropertyDao propertyDao;
+	
+	@Override
+	public Optional<String> findStringValue(String key) {
+		Optional<CmProperty> property = propertyDao.findByKeyAndParentIdAndValid(key, ROOT_ID, true);
+		return property.map(CmProperty::getValue);
+	}
+
+}