+++ server/src/main/java/com/blocklang/release/dao/AppDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blocklang.release.model.App;
 
-public interface AppDao extends JpaRepository<App, Integer>{
+public interface AppDao extends JpaRepository<App, Integer> {
 
 	Optional<App> findByRegistrationToken(String registrationToken);